# Conflicts:
#	common/src/main/java/net/mca/entity/interaction/VillagerCommandHandler.java
#	common/src/main/java/net/mca/mixin/MixinPlayerEntity.java
#	fabric/src/main/java/net/mca/fabric/MCAFabric.java
#	forge/src/main/java/net/mca/forge/ForgeBusEvents.java
#	quilt/src/main/java/net/mca/quilt/MCAQuilt.java

+++ common/src/main/java/net/mca/item/WeddingRingItem.java
@@ -12,7 +12,7 @@ public class WeddingRingItem extends RelationshipItem {
     }
 
      @Override
-    protected float getHeartsRequired() {
+    protected int getHeartsRequired() {
         return Config.getInstance().marriageHeartsRequirement;
     }
